backend/requirements-all.txt
# =============================================================================
# NOESIS Backend - Consolidated Requirements
# =============================================================================
# Install all backend dependencies: pip install -r requirements-all.txt
# =============================================================================

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.0.0
httpx>=0.27.0

# Async Support
asyncio-extras>=1.3.0
aiofiles>=24.0.0

# Database
redis>=5.0.0
qdrant-client>=1.7.0
asyncpg>=0.29.0

# LLM Providers
openai>=1.3.0

# Logging & Monitoring
structlog>=24.0.0
python-json-logger>=2.0.0

# Utils
python-dotenv>=1.0.0
pyyaml>=6.0.0
tenacity>=8.0.0

# Security
pyjwt>=2.8.0
passlib>=1.7.4
bcrypt>=4.1.0

# Testing (dev only)
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
